% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_explained_variance.R
\name{check_explained_variance}
\alias{check_explained_variance}
\title{Check explained variance}
\usage{
check_explained_variance(LD, effects, j, i)
}
\arguments{
\item{LD}{Either a centered and scaled N x P genotype matrix or a P x P matrix of correlation coefficients between the P genetic variants from an external source.}

\item{effects}{A list of three matrices, the result created by the function \code{create_effects()}.}

\item{j}{The time point.}

\item{i}{The effectiveness class; if not provided will consider all classes lumped together.}
}
\description{
Prints the amount \out{&#946;<sup>T</sup> R &#946;} of variance a set of QTL:s explain.
Here \out{&#946;} is the effect size vector of those variants, and R is their linkage disequilibrium matrix.
Also prints the sum of squared effect sizes (the amount of variance the QTL:s would explain if we disregarded LD) for comparison.
}
\seealso{
\code{\link[create_effects()]{create_effects}}
}
