% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_effects.R
\name{create_effects}
\alias{create_effects}
\title{Auxiliary function for simulating effect sizes using c distinct ''effectiveness classes''}
\usage{
create_effects(
  LD,
  amounts,
  variances,
  changes = NA,
  shuffles = NA,
  subsets = NA,
  max_reject = 100,
  prioritize_total_variance = FALSE
)
}
\arguments{
\item{LD}{Either a centered and scaled n x p genotype matrix or a p x p matrix of correlation coefficients between the p genetic variants from an external source.}

\item{amounts}{A c x t matrix of the number of QTL:s in each effectiveness class at each time point.}

\item{variances}{A c x t matrix of variances each effectiveness class explains ignoring the LD with other classes.}

\item{changes}{A c x (t - 1) matrix containing the number of QTL:s that change between time points in each effectiveness class, in addition to changes dictated by the matrix amounts.
Must have changes\link{i, j} at most min(amounts\link{i, j}, amounts\link{i, j + 1}). Not required when t = 1.}

\item{shuffles}{A c x (t - 1) logical matrix coding whether the QTL:s that are kept between time points have their effects re-randomized or not.
Not required when t = 1.}

\item{subsets}{A list of c subsets of the set {1, ..., p} to which indices of the QTL:s of each effectiveness class are restricted. By default there are no restrictions.}

\item{max_reject}{Maximum number of times the rejection sampling loop is executed. If reached, accepts the sample but gives a warning. Default value 100.}

\item{prioritize_total_variance}{If TRUE, will scale the effect sizes so that the variance explained by all the effectiveness classes jointly is the column sum of variances.
At the cost of accuracy on the variance explained by classes individually. Default value FALSE.}
}
\value{
A p x t matrix of effect sizes.
}
\description{
Auxiliary function for simulating effect sizes using c distinct ''effectiveness classes''
}
