% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_effects.R
\name{plot_effects}
\alias{plot_effects}
\title{Plot effects}
\usage{
plot_effects(effects, palette = palette.colors()[2:9])
}
\arguments{
\item{effects}{A list of three matrices, the result created by the function \code{create_effects()}.}

\item{palette}{The palette used for plotting. Default is the inclusive Okabe-Ito palette without black.}
}
\value{
Returns the ggplot.
}
\description{
Visualizes the effects of QTL:s from various effectiveness classes over time.
}
\seealso{
\code{\link[create_effects()]{create_effects}}
}
