% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_kernel.R
\name{residual_kernel}
\alias{residual_kernel}
\title{Covariance matrices for residual autocorrelation between time points}
\usage{
residual_kernel(
  variances,
  type = "noise",
  params = NA,
  points = NA,
  sample = FALSE
)
}
\arguments{
\item{variances}{A T-vector of residual variance components over time points.
This forms the diagonal of the finished covariance matrix;
first a matrix K with unit diagonal elements is created and then all rows and columns are multiplied with the corresponding standard deviation.}

\item{type}{Determines the kernel function of the Gaussian process. Available options are:
\itemize{
\item "noise" White Gaussian noise. The matrix K is the T x T identity matrix. The default option.
\item "squared" Squared exponential kernel, also known as radial basis function kernel or Gaussian kernel
The (i, j)-element of K is defined as
\ifelse{html}{\out{exp(-0.5 d(i, j)<sup>2</sup> / l<sup>2</sup> ,}}{\eqn{\exp\left(-0.5 d(i, j)^2 / l^2 \right) \,,}}
where \ifelse{html}{\out{d(i, j)}}{d(i, j)} = |\verb{points[}i\verb{]} - \verb{points[}j\verb{]}| is the distance between i:th and j:th time points
and \ifelse{html}{\out{l}}{l} = \code{params[1]} is the range parameter describing how many units away points significantly influence others.
}}

\item{params}{The parameters
\itemize{
\item parameter 1 stuff
\item parameter 2 stuff
}}
}
\description{
Returns the covariance matrix of the normally distributed residuals of a single sample over time points.
}
