% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_covariance.R
\name{stable_covariance}
\alias{stable_covariance}
\title{Stable covariance}
\usage{
stable_covariance(M)
}
\arguments{
\item{M}{An A x B data matrix.}
}
\value{
A B x B positive definite matrix that estimates the covariance matrix of the distribution that produced \code{M}.
}
\description{
The sample covariance matrices needed by various functions of the package are sometimes not positive definite, for actual or numeric reasons.
We use the functions \code{is.positive.definite()} and \code{cov.shrink()} from the package \code{corpcor} to identify and fix the issue.
}
