% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_kernel.R
\name{residual_kernel}
\alias{residual_kernel}
\title{Covariance matrices for residual autocorrelation between time points}
\usage{
residual_kernel(
  variances,
  type = "noise",
  params = NA,
  points = NA,
  sample = FALSE
)
}
\arguments{
\item{variances}{A T-vector of residual variance components over time points.
This forms the diagonal of the finished covariance matrix;
first a matrix K with unit diagonal elements is created and then all rows and columns are multiplied with the corresponding standard deviation.}

\item{type}{Determines the kernel function of the Gaussian process. Available options are:
\itemize{
\item "noise" White Gaussian noise. The matrix K is the T x T identity matrix. The default option.
\item "squared" Squared exponential kernel, also known as radial basis function kernel or Gaussian kernel
The (i, j)-element of K is defined as
\item "o-u" The Ornstein-Uhlenbeck -kernel.
}}

\item{params}{The parameters
\itemize{
\item parameter 1 stuff
\item parameter 2 stuff
}}
}
\description{
Returns the covariance matrix of the normally distributed residuals of a single sample over time points.
}
