% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_covariance.R
\name{stable_covariance}
\alias{stable_covariance}
\title{Description
The functions cov.shrink, cor.shrink, and pcor.shrink implement a shrinkage approach to estimate covariance and (partial) correlation matrices, as suggested in Schaefer and Strimmer (2005). The advantages of using this approach in comparison with the standard empirical estimates (cov and cor) are that the shrinkage estimates
are always positive definite,
well conditioned (so that the inverse always exists), and
exhibit (sometimes dramatically) better mean squared error.
Furthermore, they are inexpensive to compute and do not require any tuning parameters (the shrinkage intensity is analytically estimated from the data).
export}
\usage{
stable_covariance(M)
}
\description{
Description
The functions cov.shrink, cor.shrink, and pcor.shrink implement a shrinkage approach to estimate covariance and (partial) correlation matrices, as suggested in Schaefer and Strimmer (2005). The advantages of using this approach in comparison with the standard empirical estimates (cov and cor) are that the shrinkage estimates
are always positive definite,
well conditioned (so that the inverse always exists), and
exhibit (sometimes dramatically) better mean squared error.
Furthermore, they are inexpensive to compute and do not require any tuning parameters (the shrinkage intensity is analytically estimated from the data).
export
}
